factor out filterpopover into a shared component, reuse on Dashboard and LLM Logs page
