+++ ui/app/workspace/logs/views/filters.tsx
@@ -1,14 +1,12 @@
 import { Button } from "@/components/ui/button";
-import { Checkbox } from "@/components/ui/checkbox";
-import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList } from "@/components/ui/command";
+import { Command, CommandItem, CommandList } from "@/components/ui/command";
 import { DateTimePickerWithRange } from "@/components/ui/datePickerWithRange";
+import { FilterPopover } from "@/components/filters/filterPopover";
 import { Input } from "@/components/ui/input";
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
-import { RequestTypeLabels, RequestTypes, RoutingEngineUsedLabels, Statuses } from "@/lib/constants/logs";
-import { getErrorMessage, useGetAvailableFilterDataQuery, useGetProvidersQuery, useRecalculateLogCostsMutation } from "@/lib/store";
+import { getErrorMessage, useRecalculateLogCostsMutation } from "@/lib/store";
 import type { LogFilters as LogFiltersType } from "@/lib/types/logs";
-import { cn } from "@/lib/utils";
-import { Calculator, Check, FilterIcon, MoreVertical, Pause, Play, Search } from "lucide-react";
+import { Calculator, MoreVertical, Pause, Play, Search } from "lucide-react";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { toast } from "sonner";
 
@@ -90,7 +88,6 @@ interface LogFiltersProps {
 }
 
 export function LogFilters({ filters, onFiltersChange, liveEnabled, onLiveToggle, fetchLogs, fetchStats }: LogFiltersProps) {
-	const [openFiltersPopover, setOpenFiltersPopover] = useState(false);
 	const [openMoreActionsPopover, setOpenMoreActionsPopover] = useState(false);
 	const [localSearch, setLocalSearch] = useState(filters.content_search || "");
 	const searchTimeoutRef = useRef<NodeJS.Timeout | undefined>(undefined);
@@ -110,22 +107,6 @@ export function LogFilters({ filters, onFiltersChange, liveEnabled, onLiveToggle
 	// Convert ISO strings from filters to Date objects for the DateTimePicker
 	const [startTime, setStartTime] = useState<Date | undefined>(filters.start_time ? new Date(filters.start_time) : undefined);
 	const [endTime, setEndTime] = useState<Date | undefined>(filters.end_time ? new Date(filters.end_time) : undefined);
-
-	// Use RTK Query to fetch available models
-	const { data: providersData, isLoading: providersLoading } = useGetProvidersQuery();
-	const { data: filterData, isLoading: filterDataLoading } = useGetAvailableFilterDataQuery();
-
-	const availableProviders = providersData || [];
-	const availableModels = filterData?.models || [];
-	const availableSelectedKeys = filterData?.selected_keys || [];
-	const availableVirtualKeys = filterData?.virtual_keys || [];
-	const availableRoutingRules = filterData?.routing_rules || [];
-	const availableRoutingEngines = filterData?.routing_engines || [];
-
-	// Create mappings from name to ID for keys and virtual keys
-	const selectedKeyNameToId = new Map(availableSelectedKeys.map((key) => [key.name, key.id]));
-	const virtualKeyNameToId = new Map(availableVirtualKeys.map((key) => [key.name, key.id]));
-	const routingRuleNameToId = new Map(availableRoutingRules.map((rule) => [rule.name, rule.id]));
 
 	// Sync local date state when filters change from URL
 	useEffect(() => {
@@ -174,94 +155,12 @@ export function LogFilters({ filters, onFiltersChange, liveEnabled, onLiveToggle
 		[onFiltersChange],
 	);
 
-	const handleFilterSelect = (category: keyof typeof FILTER_OPTIONS, value: string) => {
-		const filterKeyMap: Record<keyof typeof FILTER_OPTIONS, keyof LogFiltersType> = {
-			Status: "status",
-			Providers: "providers",
-			Type: "objects",
-			Models: "models",
-			"Selected Keys": "selected_key_ids",
-			"Virtual Keys": "virtual_key_ids",
-			"Routing Rules": "routing_rule_ids",
-			"Routing Engines": "routing_engine_used",
-		};
-
-		const filterKey = filterKeyMap[category];
-		let valueToStore = value;
-
-		// Convert name to ID for keys and virtual keys
-		if (category === "Selected Keys") {
-			valueToStore = selectedKeyNameToId.get(value) || value;
-		} else if (category === "Virtual Keys") {
-			valueToStore = virtualKeyNameToId.get(value) || value;
-		} else if (category === "Routing Rules") {
-			valueToStore = routingRuleNameToId.get(value) || value;
-		}
-
-		const currentValues = (filters[filterKey] as string[]) || [];
-		const newValues = currentValues.includes(valueToStore)
-			? currentValues.filter((v) => v !== valueToStore)
-			: [...currentValues, valueToStore];
-
-		onFiltersChange({
-			...filters,
-			[filterKey]: newValues,
-		});
-	};
-
-	const isSelected = (category: keyof typeof FILTER_OPTIONS, value: string) => {
-		const filterKeyMap: Record<keyof typeof FILTER_OPTIONS, keyof LogFiltersType> = {
-			Status: "status",
-			Providers: "providers",
-			Type: "objects",
-			Models: "models",
-			"Selected Keys": "selected_key_ids",
-			"Virtual Keys": "virtual_key_ids",
-			"Routing Rules": "routing_rule_ids",
-			"Routing Engines": "routing_engine_used",
-		};
-
-		const filterKey = filterKeyMap[category];
-		const currentValues = filters[filterKey];
-
-		// For keys and virtual keys, convert name to ID before checking
-		let valueToCheck = value;
-		if (category === "Selected Keys") {
-			valueToCheck = selectedKeyNameToId.get(value) || value;
-		} else if (category === "Virtual Keys") {
-			valueToCheck = virtualKeyNameToId.get(value) || value;
-		} else if (category === "Routing Rules") {
-			valueToCheck = routingRuleNameToId.get(value) || value;
-		}
-
-		return Array.isArray(currentValues) && currentValues.includes(valueToCheck);
-	};
-
-	const getSelectedCount = () => {
-		// Exclude timestamp filters and content_search from the count
-		const excludedKeys = ["start_time", "end_time", "content_search"];
-
-		return Object.entries(filters).reduce((count, [key, value]) => {
-			if (excludedKeys.includes(key)) {
-				return count;
-			}
-			if (Array.isArray(value)) {
-				return count + value.length;
-			}
-			return count + (value ? 1 : 0);
-		}, 0);
-	};
-
-	const FILTER_OPTIONS = {
-		Status: Statuses,
-		Providers: providersLoading ? ["Loading providers..."] : availableProviders.map((provider) => provider.name),
-		Type: RequestTypes,
-		Models: filterDataLoading ? ["Loading models..."] : availableModels,
-		"Selected Keys": filterDataLoading ? ["Loading selected keys..."] : availableSelectedKeys.map((key) => key.name),
-		"Virtual Keys": filterDataLoading ? ["Loading virtual keys..."] : availableVirtualKeys.map((key) => key.name),
-		"Routing Engines": filterDataLoading ? ["Loading routing engines..."] : availableRoutingEngines,
-		"Routing Rules": filterDataLoading ? ["Loading routing rules..."] : availableRoutingRules.map((rule) => rule.name),
-	} as const;
+	const handleFilterChange = useCallback(
+		(key: keyof LogFiltersType, values: string[] | boolean) => {
+			onFiltersChange({ ...filters, [key]: values });
+		},
+		[filters, onFiltersChange],
+	);
 
 	return (
 		<div className="flex items-center justify-between space-x-2">
@@ -317,81 +216,7 @@ export function LogFilters({ filters, onFiltersChange, liveEnabled, onLiveToggle
 					});
 				}}
 			/>
-			<Popover open={openFiltersPopover} onOpenChange={setOpenFiltersPopover}>
-				<PopoverTrigger asChild>
-					<Button variant="outline" size="sm" className="h-7.5 w-[120px]">
-						<FilterIcon className="h-4 w-4" />
-						Filters
-						{getSelectedCount() > 0 && (
-							<span className="bg-primary/10 flex h-6 w-6 items-center justify-center rounded-full text-xs font-normal">
-								{getSelectedCount()}
-							</span>
-						)}
-					</Button>
-				</PopoverTrigger>
-				<PopoverContent className="w-[200px] p-0" align="end">
-					<Command>
-						<CommandInput placeholder="Search filters..." />
-						<CommandList>
-							<CommandEmpty>No filters found.</CommandEmpty>
-							<CommandGroup>
-								<CommandItem className="cursor-pointer">
-									<Checkbox
-										className={cn(
-											"border-primary opacity-50",
-											filters.missing_cost_only && "bg-primary text-primary-foreground opacity-100",
-										)}
-										id="missing-cost-toggle"
-										checked={!!filters.missing_cost_only}
-										onCheckedChange={(checked: boolean) => onFiltersChange({ ...filters, missing_cost_only: checked })}
-									/>
-									<span className="text-sm">Show missing cost</span>
-								</CommandItem>
-							</CommandGroup>
-							{Object.entries(FILTER_OPTIONS)
-								.filter(([_, values]) => values.length > 0)
-								.map(([category, values]) => (
-									<CommandGroup key={category} heading={category}>
-										{values.map((value: string) => {
-											const selected = isSelected(category as keyof typeof FILTER_OPTIONS, value);
-											const isLoading =
-												(category === "Providers" && providersLoading) ||
-												(category === "Models" && filterDataLoading) ||
-												(category === "Selected Keys" && filterDataLoading) ||
-												(category === "Virtual Keys" && filterDataLoading) ||
-												(category === "Routing Rules" && filterDataLoading) ||
-												(category === "Routing Engines" && filterDataLoading);
-											return (
-												<CommandItem
-													key={value}
-													onSelect={() => !isLoading && handleFilterSelect(category as keyof typeof FILTER_OPTIONS, value)}
-													disabled={isLoading}
-												>
-													<div
-														className={cn(
-															"border-primary mr-2 flex h-4 w-4 items-center justify-center rounded-sm border",
-															selected ? "bg-primary text-primary-foreground" : "opacity-50 [&_svg]:invisible",
-														)}
-													>
-														{isLoading ? (
-															<div className="border-primary h-3 w-3 animate-spin rounded-full border border-t-transparent" />
-														) : (
-															<Check className="text-primary-foreground size-3" />
-														)}
-													</div>
-													<span className={cn("lowercase", isLoading && "text-muted-foreground")}>
-														{category === "Type" ? RequestTypeLabels[value as keyof typeof RequestTypeLabels] :
-															category === "Routing Engines" ? (RoutingEngineUsedLabels[value as keyof typeof RoutingEngineUsedLabels] ?? value) : value}
-													</span>
-												</CommandItem>
-											);
-										})}
-									</CommandGroup>
-								))}
-						</CommandList>
-					</Command>
-				</PopoverContent>
-			</Popover>
+			<FilterPopover filters={filters} onFilterChange={handleFilterChange} showMissingCost />
 			<Popover open={openMoreActionsPopover} onOpenChange={setOpenMoreActionsPopover}>
 				<PopoverTrigger asChild>
 					<Button variant="outline" size="sm" className="h-7.5">
